fix dataSources store type
