+++ packages/react-redux/src/slices/cartoSlice.js
@@ -206,7 +206,7 @@ export const createCartoSlice = (initialState) => {
  * @param {string} data.type - type of source. Possible values are sql or bigquery.
  * @param {object=} data.credentials - (optional) Custom credentials to be used in the source.
  * @param {string} data.connection - connection name for CARTO 3 source.
- * @param {import('./cartoSlice').SourceFilters=} data.filters - logical operator that defines how filters for different columns are joined together.
+ * @param {import('../types').SourceFilters=} data.filters - logical operator that defines how filters for different columns are joined together.
  * @param {FiltersLogicalOperators=} data.filtersLogicalOperator - logical operator that defines how filters for different columns are joined together.
  * @param {import('@deck.gl/carto/typed').QueryParameters} data.queryParameters - SQL query parameters.
  * @param {string=} data.geoColumn - (optional) name of column containing geometries or spatial index data.
